+++ src/Appwrite/Platform/Modules/Databases/Http/Grids/Tables/Delete.php
@@ -34,7 +34,7 @@ class Delete extends CollectionDelete
             ->label('resourceType', RESOURCE_TYPE_DATABASES)
             ->label('event', 'databases.[databaseId].tables.[tableId].delete')
             ->label('audits.event', 'table.delete')
-            ->label('audits.resource', 'database/{request.databaseId}/table/{request.tableId}')
+            ->label('audits.resource', 'database/{request.databaseId}/grid/table/{request.tableId}')
             ->label('sdk', new Method(
                 namespace: $this->getSdkNamespace(),
                 group: null,