+++ tests/e2e/Services/Avatars/AvatarsBase.php
@@ -284,7 +284,17 @@ trait AvatarsBase
         ]);
 
         $this->assertEquals(200, $response['headers']['status-code']);
-        $this->assertEquals('image/x-icon', $response['headers']['content-type']);
+        $this->assertEquals('image/svg+xml', $response['headers']['content-type']);
+        $this->assertNotEmpty($response['body']);
+
+        $response = $this->client->call(Client::METHOD_GET, '/avatars/favicon', [
+            'x-appwrite-project' => $this->getProject()['$id'],
+        ], [
+            'url' => 'https://appwrite.io/',
+        ]);
+
+        $this->assertEquals(200, $response['headers']['status-code']);
+        $this->assertEquals('image/svg+xml', $response['headers']['content-type']);
         $this->assertNotEmpty($response['body']);
 
         /**
@@ -337,7 +347,7 @@ trait AvatarsBase
         $this->assertEquals(400, $image->getImageWidth());
         $this->assertEquals(400, $image->getImageHeight());
         $this->assertEquals('PNG', $image->getImageFormat());
-        $this->assertEquals(strlen(\file_get_contents(__DIR__ . '/../../../resources/qr/qr-default.png')), strlen($response['body']));
+        $this->assertSamePixels(__DIR__ . '/../../../resources/qr/qr-default.png', $response['body']);
 
         $response = $this->client->call(Client::METHOD_GET, '/avatars/qr', [
             'x-appwrite-project' => $this->getProject()['$id'],
@@ -355,7 +365,7 @@ trait AvatarsBase
         $this->assertEquals(200, $image->getImageWidth());
         $this->assertEquals(200, $image->getImageHeight());
         $this->assertEquals('PNG', $image->getImageFormat());
-        $this->assertEquals(strlen(\file_get_contents(__DIR__ . '/../../../resources/qr/qr-size-200.png')), strlen($response['body']));
+        $this->assertSamePixels(__DIR__ . '/../../../resources/qr/qr-size-200.png', $response['body']);
 
         $response = $this->client->call(Client::METHOD_GET, '/avatars/qr', [
             'x-appwrite-project' => $this->getProject()['$id'],
@@ -374,7 +384,7 @@ trait AvatarsBase
         $this->assertEquals(200, $image->getImageWidth());
         $this->assertEquals(200, $image->getImageHeight());
         $this->assertEquals('PNG', $image->getImageFormat());
-        $this->assertEquals(strlen(\file_get_contents(__DIR__ . '/../../../resources/qr/qr-size-200-margin-10.png')), strlen($response['body']));
+        $this->assertSamePixels(__DIR__ . '/../../../resources/qr/qr-size-200-margin-10.png', $response['body']);
 
         $response = $this->client->call(Client::METHOD_GET, '/avatars/qr', [
             'x-appwrite-project' => $this->getProject()['$id'],
@@ -390,7 +400,7 @@ trait AvatarsBase
         $this->assertEquals(200, $image->getImageWidth());
         $this->assertEquals(200, $image->getImageHeight());
         $this->assertEquals('PNG', $image->getImageFormat());
-        $this->assertEquals(strlen(\file_get_contents(__DIR__ . '/../../../resources/qr/qr-size-200-margin-10.png')), strlen($response['body']));
+        $this->assertSamePixels(__DIR__ . '/../../../resources/qr/qr-size-200-margin-10.png', $response['body']);
 
         $this->assertEquals(200, $response['headers']['status-code']);
         $this->assertEquals('attachment; filename="qr.png"', $response['headers']['content-disposition']);