+++ tests/e2e/Services/Databases/Legacy/DatabasesBase.php
@@ -32,6 +32,44 @@ trait DatabasesBase
         $this->assertNotEmpty($database['body']['$id']);
         $this->assertEquals(201, $database['headers']['status-code']);
         $this->assertEquals('Test Database', $database['body']['name']);
+        $this->assertEquals('grids', $database['body']['type']);
+
+        // testing to create a database with type
+        $database2 = $this->client->call(Client::METHOD_POST, '/databases', [
+            'content-type' => 'application/json',
+            'x-appwrite-project' => $this->getProject()['$id'],
+            'x-appwrite-key' => $this->getProject()['apiKey']
+        ], [
+            'databaseId' => ID::unique(),
+            'name' => 'Test Database with type',
+            'type' => 'mongodb'
+        ]);
+        $this->assertEquals(400, $database2['headers']['status-code']);
+
+        $database2 = $this->client->call(Client::METHOD_POST, '/databases', [
+            'content-type' => 'application/json',
+            'x-appwrite-project' => $this->getProject()['$id'],
+            'x-appwrite-key' => $this->getProject()['apiKey']
+        ], [
+            'databaseId' => ID::unique(),
+            'name' => 'Test Database with type',
+            'type' => 'legacy'
+        ]);
+        $this->assertNotEmpty($database2['body']['$id']);
+        $this->assertEquals(201, $database2['headers']['status-code']);
+        $this->assertEquals('Test Database with type', $database2['body']['name']);
+        $this->assertEquals('legacy', $database2['body']['type']);
+
+        // cleanup(for database2)
+        $databaseId = $database2['body']['$id'];
+
+        $response = $this->client->call(Client::METHOD_DELETE, '/databases/' . $databaseId, [
+            'content-type' => 'application/json',
+            'x-appwrite-project' => $this->getProject()['$id'],
+            'x-appwrite-key' => $this->getProject()['apiKey']
+        ]);
+
+        $this->assertEquals(204, $response['headers']['status-code']);
 
         return ['databaseId' => $database['body']['$id']];
     }
@@ -1639,6 +1677,7 @@ trait DatabasesBase
         $this->assertEquals($document1['body']['actors'][1], 'Samuel Jackson');
         $this->assertEquals($document1['body']['birthDay'], '1975-06-12T12:12:55.000+00:00');
         $this->assertTrue(array_key_exists('$sequence', $document1['body']));
+        $this->assertIsInt($document1['body']['$sequence']);
 
         $this->assertEquals(201, $document2['headers']['status-code']);
         $this->assertEquals($data['moviesId'], $document2['body']['$collectionId']);
@@ -2876,7 +2915,6 @@ trait DatabasesBase
                 'releaseYear' => 2017,
                 'birthDay' => '1976-06-12 14:12:55',
                 'actors' => [],
-                '$createdAt' => 5 // Should be ignored
             ],
             'permissions' => [
                 Permission::read(Role::user($this->getUser()['$id'])),
@@ -4282,16 +4320,20 @@ trait DatabasesBase
         $document = $this->client->call(Client::METHOD_PATCH, '/databases/' . $databaseId . '/collections/' . $moviesId . '/documents/' . $documentId, $headers, [
             'data' => [
                 'title' => 'Again Updated Date Test',
-                '$createdAt' => '2022-08-01 13:09:23.040', // $createdAt is not updatable
-                '$updatedAt' => '2022-08-01 13:09:23.050' // system will update it not api
+                '$createdAt' => '2022-08-01 13:09:23.040',
+                '$updatedAt' => '2022-08-01 13:09:23.050'
             ]
         ]);
-        $this->assertEquals($document['body']['title'], 'Again Updated Date Test');
-        $this->assertEquals($document['body']['$createdAt'], $createdAt);
-        $this->assertNotEquals($document['body']['$createdAt'], '2022-08-01 13:09:23.040');
-        $this->assertNotEquals($document['body']['$updatedAt'], $updatedAt);
-        $this->assertNotEquals($document['body']['$updatedAt'], $updatedAtSecond);
-        $this->assertNotEquals($document['body']['$updatedAt'], '2022-08-01 13:09:23.050');
+        if ($this->getSide() === 'client') {
+            $this->assertEquals($document['headers']['status-code'], 400);
+        } else {
+            $this->assertEquals($document['body']['title'], 'Again Updated Date Test');
+            $this->assertEquals($document['body']['$createdAt'], DateTime::formatTz('2022-08-01 13:09:23.040'));
+            $this->assertEquals($document['body']['$updatedAt'], DateTime::formatTz('2022-08-01 13:09:23.050'));
+
+        }
+
+        return $data;
     }
 
     public function testUpdatePermissionsWithEmptyPayload(): array
@@ -5770,6 +5812,26 @@ trait DatabasesBase
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ]), ['min' => 7]);
+        $this->assertEquals(400, $err['headers']['status-code']);
+
+        // Test min limit exceeded with custom value
+        $err = $this->client->call(Client::METHOD_PATCH, '/databases/' . $databaseId . '/collections/' . $collectionId . '/documents/' . $documentId . '/count/decrement', array_merge([
+            'content-type' => 'application/json',
+            'x-appwrite-project' => $this->getProject()['$id'],
+        ]), [
+            'value' => 3,
+            'min' => 5,
+        ]);
+        $this->assertEquals(400, $err['headers']['status-code']);
+
+        // Test min limit 0
+        $err = $this->client->call(Client::METHOD_PATCH, '/databases/' . $databaseId . '/collections/' . $collectionId . '/documents/' . $documentId . '/count/decrement', array_merge([
+            'content-type' => 'application/json',
+            'x-appwrite-project' => $this->getProject()['$id'],
+        ]), [
+            'value' => 10,
+            'min' => 0,
+        ]);
         $this->assertEquals(400, $err['headers']['status-code']);
 
         // Test type error on non-numeric attribute