+++ docs/sdks/php/CHANGELOG.md
@@ -1,1 +1,9 @@
 # Change Log
+
+## 15.1.0
+
+* Add `incrementDocumentAttribute` and `decrementDocumentAttribute` support to `Databases` service
+* Add `dart38` and `flutter332` support to runtime models
+* Add `gif` support to `ImageFormat` enum
+* Add `upsertDocument` support to `Databases` service
+* Add `sequence` support to `Document` model